+++ chuffed/examples/graceful_graph.cpp
@@ -46,7 +46,7 @@ public:
 			}
 		}
 
-		assert(origins.size() == nedges);
+		assert(static_cast<int>(origins.size()) == nedges);
 
 		createVars(x, nodes, 0, nedges);
 		createVars(d, nedges, -nedges, nedges);
@@ -118,7 +118,7 @@ public:
 		for (int i = 0; i < sat.nVars(); i++) {
 			sat.flags[i] = LitFlags(false, false, false);
 		}
-		for (int i = 0; i < x.size(); i++) {
+		for (unsigned int i = 0; i < x.size(); i++) {
 			assert(x[i]->getType() == INT_VAR_EL);
 			((IntVarEL*)x[i])->setVLearnable();
 			((IntVarEL*)x[i])->setVDecidable(true);
